Change `withCache` to never send values that could be stale.
When an observable created with `withCache` has observers, it will store
the latest version of the underlying observable so that new observers see it
immediately. When all observers unobserve, the `withCache` also unobserves
the underlying observable.

Before this change, `withCache` did not reset its internal state when unobserving
the underlying observable. This means that if it was observed, then observed down to
0 observers, then the underlying observable updated, it would send a stale value
to any new observers before getting the correct state from the underlying observable.
This is particularly counter intuitive for observers like `toPromise` that will
unsubscribe themselves after a single value comes through.

This changes clear the internal state when `withCache` becomes completely unobserved.
This helps guarantee that a stale state will never be sent through to observers.